backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
PyMySQL==1.1.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
eventlet==0.35.1
